Delete unnecessary Spline3D component
