+++ src/components/home/MainHeader.tsx
@@ -1,9 +1,13 @@
+'use client';
+
 import { CTA } from '@/components/CTA';
 import { HeadTitle } from '@/components/HeadTitle';
 import { SocialMedia } from '@/components/SocialMedia';
-import { Spline3D } from '@/components/home/Spline3D';
+import Spinner from "@/components/Spinner";
 import Image from 'next/image';
+import { lazy, Suspense } from 'react';
 import { getAge } from 'src/utils';
+const Spline = lazy(() => import('@splinetool/react-spline'));
 
 
 export const MainHeader = () => (
@@ -37,7 +41,13 @@ export const MainHeader = () => (
             </div>
         </div>
         <div className='mt-10 overflow-hidden rounded-xl'>
-            <Spline3D />
+            <Suspense fallback={(
+                <div className="hidden xl:visible xl:grid xl:place-content-center">
+                    <Spinner />
+                </div>
+            )}>
+                <Spline scene="https://prod.spline.design/FFs-WdvYFpj7feEm/scene.splinecode"  />
+            </Suspense>
         </div>
     </header>
 );
